reorg(#2514): fix depth-shifted relative paths after the lib relocation. Pattern 1 (add one upward level): the 3 seed-bin dynamic imports, config eslint/vite/test repo-root reaches, rules-test setup, and the 4 intentionsutil scripts' repoRoot computation. Pattern 2 (insert packages/): the 3 blog refs from fellspiral and functions. Includes 5 references beyond the plan's non-exhaustive anchor list, found via the plan-mandated sweep (config/vite.js envDir + 4 intentionsutil repoRoot).
